move date related util function to a single file
